--- a/Makefile
+++ b/Makefile
@@ -9,17 +9,9 @@ TSCRIPT = psutil/tests/runner.py
 
 # Internal.
 PY3_DEPS = \
-	autoflake \
-	autopep8 \
 	check-manifest \
 	concurrencytest \
 	coverage \
-	flake8 \
-	flake8-blind-except \
-	flake8-bugbear \
-	flake8-debugger \
-	flake8-print \
-	flake8-quotes \
 	isort \
 	pep8-naming \
 	pylint \
@@ -200,8 +192,8 @@ test-coverage:  ## Run test coverage.
 # Linters
 # ===================================================================
 
-flake8:  ## Run flake8 linter.
-	@git ls-files '*.py' | xargs $(PYTHON) -m flake8 --config=.flake8 --jobs=${NUM_WORKERS}
+ruff:  ## Run ruff linter.
+	@git ls-files '*.py' | xargs $(PYTHON) -m ruff check --config=pyproject.toml --no-cache
 
 isort:  ## Run isort linter.
 	@git ls-files '*.py' | xargs $(PYTHON) -m isort --check-only --jobs=${NUM_WORKERS}
@@ -219,7 +211,6 @@ lint-toml:  ## Linter for pyproject.toml
 	@git ls-files '*.toml' | xargs toml-sort --check
 
 lint-all:  ## Run all linters
-	${MAKE} flake8
 	${MAKE} isort
 	${MAKE} lint-c
 	${MAKE} lint-rst
@@ -229,9 +220,8 @@ lint-all:  ## Run all linters
 # Fixers
 # ===================================================================
 
-fix-flake8:  ## Run autopep8, fix some Python flake8 / pep8 issues.
-	@git ls-files '*.py' | xargs $(PYTHON) -m autopep8 --in-place --jobs=${NUM_WORKERS} --global-config=.flake8
-	@git ls-files '*.py' | xargs $(PYTHON) -m autoflake --in-place --jobs=${NUM_WORKERS} --remove-all-unused-imports --remove-unused-variables --remove-duplicate-keys
+fix-ruff:
+	@git ls-files '*.py' | xargs $(PYTHON) -m ruff --config=pyproject.toml --no-cache --fix
 
 fix-imports:  ## Fix imports with isort.
 	@git ls-files '*.py' | xargs $(PYTHON) -m isort --jobs=${NUM_WORKERS}
@@ -243,7 +233,7 @@ fix-toml:  ## Fix pyproject.toml
 	@git ls-files '*.toml' | xargs toml-sort
 
 fix-all:  ## Run all code fixers.
-	${MAKE} fix-flake8
+	${MAKE} fix-ruff
 	${MAKE} fix-imports
 	${MAKE} fix-unittests
 	${MAKE} fix-toml