--- a/docs/conf.py
+++ b/docs/conf.py
@@ -34,7 +34,7 @@ HERE = os.path.abspath(os.path.dirname(__file__))
 
 def get_version():
     INIT = os.path.abspath(os.path.join(HERE, '../psutil/__init__.py'))
-    with open(INIT, 'r') as f:
+    with open(INIT) as f:
         for line in f:
             if line.startswith('__version__'):
                 ret = eval(line.strip().split(' = ')[1])