--- a/psutil/__init__.py
+++ b/psutil/__init__.py
@@ -227,8 +227,8 @@ _SENTINEL = object()
 # See: https://github.com/giampaolo/psutil/issues/564
 if (int(__version__.replace('.', '')) !=
         getattr(_psplatform.cext, 'version', None)):
-    msg = "version conflict: %r C extension module was built for another " \
-          "version of psutil" % _psplatform.cext.__file__
+    msg = "version conflict: %r C extension module was built for another " + \
+        "version of psutil" % _psplatform.cext.__file__
     if hasattr(_psplatform.cext, 'version'):
         msg += " (%s instead of %s)" % (
             '.'.join([x for x in str(_psplatform.cext.version)]), __version__)