--- a/psutil/_common.py
+++ b/psutil/_common.py
@@ -37,7 +37,7 @@ except ImportError:
 
 
 # can't take it from _common.py as this script is imported by setup.py
-PY3 = sys.version_info[0] == 3
+PY3 = sys.version_info[0] >= 3
 if PY3:
     import enum
 else: