--- a/psutil/_compat.py
+++ b/psutil/_compat.py
@@ -222,7 +222,7 @@ except ImportError:
         "CacheInfo", ["hits", "misses", "maxsize", "currsize"])
 
     class _HashedSeq(list):
-        __slots__ = 'hashvalue'
+        __slots__ = ('hashvalue', )
 
         def __init__(self, tup, hash=hash):
             self[:] = tup