--- a/psutil/_psbsd.py
+++ b/psutil/_psbsd.py
@@ -226,14 +226,14 @@ def swap_memory():
 
 
 def cpu_times():
-    """Return system per-CPU times as a namedtuple"""
+    """Return system per-CPU times as a namedtuple."""
     user, nice, system, idle, irq = cext.cpu_times()
     return scputimes(user, nice, system, idle, irq)
 
 
 if HAS_PER_CPU_TIMES:
     def per_cpu_times():
-        """Return system CPU times as a namedtuple"""
+        """Return system CPU times as a namedtuple."""
         ret = []
         for cpu_t in cext.per_cpu_times():
             user, nice, system, idle, irq = cpu_t
@@ -249,7 +249,7 @@ else:
     # crash at psutil import time.
     # Next calls will fail with NotImplementedError
     def per_cpu_times():
-        """Return system CPU times as a namedtuple"""
+        """Return system CPU times as a namedtuple."""
         if cpu_count_logical() == 1:
             return [cpu_times()]
         if per_cpu_times.__called__:
@@ -365,7 +365,7 @@ elif OPENBSD:
 def disk_partitions(all=False):
     """Return mounted disk partitions as a list of namedtuples.
     'all' argument is ignored, see:
-    https://github.com/giampaolo/psutil/issues/906
+    https://github.com/giampaolo/psutil/issues/906.
     """
     retlist = []
     partitions = cext.disk_partitions()