--- a/psutil/_pslinux.py
+++ b/psutil/_pslinux.py
@@ -1191,6 +1191,7 @@ class RootFsDeviceFinder:
     https://bootlin.com/blog/find-root-device/
     https://www.systutorials.com/how-to-find-the-disk-where-root-is-on-in-bash-on-linux/
     """
+
     __slots__ = ['major', 'minor']
 
     def __init__(self):