--- a/psutil/_psosx.py
+++ b/psutil/_psosx.py
@@ -354,7 +354,7 @@ def wrap_exceptions(fun):
     return wrapper
 
 
-class Process(object):
+class Process:
     """Wrapper class around underlying C implementation."""
 
     __slots__ = ["pid", "_name", "_ppid", "_cache"]