--- a/psutil/_psposix.py
+++ b/psutil/_psposix.py
@@ -78,7 +78,7 @@ else:  # pragma: no cover
 
 def wait_pid(pid, timeout=None, proc_name=None,
              _waitpid=os.waitpid,
-             _timer=getattr(time, 'monotonic', time.time),
+             _timer=getattr(time, 'monotonic', time.time),  # noqa: B008
              _min=min,
              _sleep=time.sleep,
              _pid_exists=pid_exists):