See: https://blog.jerrycodes.com/ruff-the-python-linter/. Advantages:

1) an order of magnitude faster than flake8 + isort (0.1 secs instead of 2.5 secs to lint all .py files):

Before:
```
$ time make flake8 isort 
real	0m2,554s
user	0m15,626s
sys	0m0,436s
```

Now:
```
$ time make ruff 
real	0m0,102s
user	0m0,297s
sys	0m0,025s
```

2) A lot more code quality checks than before.

Previously done also for pyftpdlib: https://github.com/giampaolo/pyftpdlib/pull/611.